requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
google-cloud-vision==3.8.1
requests==2.32.3
streamlit==1.40.2
pillow==11.0.0
pandas==2.2.3
plotly==5.15.0
